Licence update - modified the original LGPLv3
Additional permissions to make the licence much more permissive for Grepton Zrt. which enabled me to use this part of our repository that I was basically writing by myself at home to make our projects be able to show up 3D models in the c++ projects. This code contains ideas from the original java version, but that version is not open sourced at all as it is much more comprehensible and not only mine.

The new licence is the same as LGPLv3, just I made a change so that my corporation can use it without ANY restrictions, fork it in
closed source and everything related. This is how it is rightous to do and how we agreed.