Update README.md
Added design considerations about being to-the-point and lightweight opposed to hande various crazy edge-cases and make the code complexity grow because of the really unnecessary stuff. If the code is small enough, maybe you can look into that and see what it can load and what it can't. This is much better than guessing as there will be always something not covered I think.